祭祀项目进度，我的钱包和余额变动

+++ src/main/java/com/ruoyi/project/business/service/FamilyMemberService.java
@@ -5,11 +5,11 @@ import com.ruoyi.project.business.domain.FamilyMember;
 import com.baomidou.mybatisplus.extension.service.IService;
 public interface FamilyMemberService extends IService<FamilyMember>{
     
-    IPage<FamilyMember> selectAllFamilyMembersByTeamId(Integer pageNo, Integer pageSize, Long teamId);
+    IPage<FamilyMember> selectAllFamilyMembersByTeamId(Integer pageNo, Integer pageSize, Integer teamId);
 
     String updateFamilyMemberStatus(Integer id, Integer status);
 
-    String transFamilyMemberRole(Long teamId, Long oldUserId, Long newUserId);
+    String transFamilyMemberRole(Integer teamId, Integer oldUserId, Integer newUserId);
 
-    String deleteFamilyMember(Long id);
+    String deleteFamilyMember(Integer id);
 }
